fix problem with conversation
