+++ src/bot.ts
@@ -21,12 +21,27 @@ import {
   adminUsersMenu,
 } from '@bot/menu';
 import { auth, normalize } from '@bot/middlewares';
-import { mongo } from '@bot/services';
+import { logger, mongo } from '@bot/services';
 import { BotContextType } from '@bot/types';
 import { conversations, createConversation } from '@grammyjs/conversations';
 import { I18n } from '@grammyjs/i18n';
-import { Bot, session } from 'grammy';
+import { Bot, BotError, GrammyError, HttpError, session } from 'grammy';
 import path from 'path';
+
+const handleBotError = (error: BotError) => {
+  const ctx = error.ctx;
+  const err = error.error;
+
+  logger.error(`botInitialize::error while handling update::${ctx.update.update_id}:`);
+
+  if (err instanceof GrammyError) {
+    logger.error(`botInitialize::error in request::${err.description}`);
+  } else if (err instanceof HttpError) {
+    logger.error(`botInitialize::could not contact Telegram::${err.message}`);
+  } else {
+    logger.error(`botInitialize::unknown error::${(err as Error).message}`);
+  }
+};
 
 export const createBot = async () => {
   const bot = new Bot<BotContextType>(config.TELEGRAM_TOKEN);
@@ -78,6 +93,8 @@ export const createBot = async () => {
     textController,
     voiceController,
   ].forEach((handle) => handle(bot));
+
+  bot.catch(handleBotError);
 
   return bot;
 };
